requirements.txt
scikit-learn
streamlit
pandas
xgboost
joblib
python-dotenv
langchain
langchain-google-genai
plotly